add IE11 support for Svelte2Wrapper (by not using web components here)
