+++ src/routes/lib.js
@@ -83,11 +83,12 @@ module.exports = {
                     const { file } = request.params;
                     const { anonymous } = request.query;
                     const isIE = file.endsWith('.ie.js');
-                    const isJS = file.endsWith('.js');
-                    const isSvelte = file.includes('.svelte.js');
-                    const isJSMap = file.endsWith('.js.map');
+                    const file2 = isIE ? file.replace('.ie.js', '.js') : file;
+                    const isJS = file2.endsWith('.js');
+                    const isSvelte = file2.includes('.svelte.js');
+                    const isJSMap = file2.endsWith('.js.map');
                     const page = isSvelte
-                        ? file.replace(/\.svelte(\.ie)?\.js(\.map)?/, '.svelte')
+                        ? file2.replace(/\.svelte(\.ie)?\.js(\.map)?/, '.svelte')
                         : isJS
                         ? file
                         : `${file}.js`;
