+++ src/utils/svelte-view/template.ejs
@@ -39,6 +39,7 @@
         <script>window.__DW_SVELTE_PROPS__ = { data: { polyfillUri: '/lib/polyfills' }}</script>
         <script src="/lib/chart-core/load-polyfills.js"></script>
         <script src="/lib/static/js/svg4everybody.js"></script>
+        <script>if (window.document.documentMode) { svg4everybody(); }</script>
 
         <script type="text/javascript" src="/lib/requirejs/require.js"></script>
         <script type="text/javascript">
